Option to disable elevation of `Button`
- Also move types under component folder
