+++ packages/ui/src/components/Forms/Button/Types.ts
@@ -1,0 +1,8 @@
+import { Default } from '../../../types'
+
+export type ButtonFlow = Default | 'circular' | 'squarish' | 'generous' | 'block' | 'generousBlock'
+export type ButtonDistinction = Default | 'primary' | 'toned' | 'outlined' | 'seamless'
+export type ButtonElevation = Default | 'none'
+export type ButtonGroupFlow = Default | 'block'
+export type ButtonGroupOrientation = Default | 'horizontal' | 'vertical'
+export type ButtonListFlow = Default | 'inline' | 'block'
