+++ packages/ui/src/components/Forms/Button/index.ts
@@ -1,3 +1,4 @@
 export * from './Button'
 export * from './ButtonGroup'
 export * from './ButtonList'
+export * from './Types'
