--- packages/ui/src/types/ButtonDistinction.ts
@@ -1,3 +1,0 @@
-import type { Default } from './Default'
-
-export type ButtonDistinction = Default | 'primary' | 'toned' | 'outlined' | 'seamless'
