--- packages/ui/src/types/ButtonFlow.ts
@@ -1,3 +1,0 @@
-import type { Default } from './Default'
-
-export type ButtonFlow = Default | 'circular' | 'squarish' | 'generous' | 'block' | 'generousBlock'
