--- packages/ui/src/types/ButtonGroupFlow.ts
@@ -1,3 +1,0 @@
-import type { Default } from './Default'
-
-export type ButtonGroupFlow = Default | 'block'
