--- packages/ui/src/types/ButtonGroupOrientation.ts
@@ -1,3 +1,0 @@
-import type { Default } from './Default'
-
-export type ButtonGroupOrientation = Default | 'horizontal' | 'vertical'
