--- packages/ui/src/types/ButtonListFlow.ts
@@ -1,3 +1,0 @@
-import type { Default } from './Default'
-
-export type ButtonListFlow = Default | 'inline' | 'block'
