+++ packages/ui/src/types/index.ts
@@ -1,11 +1,6 @@
 export * from './Alignment'
 export * from './BoxDepth'
 export * from './BoxDistinction'
-export * from './ButtonDistinction'
-export * from './ButtonFlow'
-export * from './ButtonGroupFlow'
-export * from './ButtonGroupOrientation'
-export * from './ButtonListFlow'
 export * from './CollapsibleTransition'
 export * from './ControlDistinction'
 export * from './Default'
